third_party/WebKit/LayoutTests/accessibility/listbox-enabled-states-expected.txt


This tests that we report the correct enabled state on a listbox.


PASS axSelectElement.restriction is "none"
PASS axSelectElement.childAtIndex(0).restriction is "none"
PASS axSelectElement.childAtIndex(1).restriction is "none"
PASS axSelectElement.childAtIndex(2).restriction is "disabled"
PASS axSelectElement.childAtIndex(3).restriction is "disabled"
PASS axSelectElement.childAtIndex(4).restriction is "none"
PASS successfullyParsed is true

TEST COMPLETE
